+++ packages/cli/src/scripts/settlement/deploy-token.ts
@@ -4,7 +4,11 @@ import { DispatchSmartContract } from "@proto-kit/protocol";
 import "reflect-metadata";
 import { container } from "tsyringe";
 
-import { loadEnvironmentVariables, LoadEnvOptions } from "../../utils/loadEnv";
+import {
+  getRequiredEnv,
+  loadEnvironmentVariables,
+  LoadEnvOptions,
+} from "../../utils/loadEnv";
 import { loadUserModules } from "../../utils/loadUserModules";
 
 export interface TokenDeployArgs {
@@ -25,6 +29,8 @@ export default async function (
       );
     }
     loadEnvironmentVariables(options);
+    const { scriptModules, scriptModulesConfig } =
+      await import("../../utils/modules");
 
     const { Runtime } = await import("@proto-kit/module");
     const { Protocol } = await import("@proto-kit/protocol");
@@ -37,6 +43,7 @@ export default async function (
       SignedSettlementPermissions,
       AppChain,
       BridgingModule,
+      InMemoryDatabase,
     } = await import("@proto-kit/sequencer");
     const {
       AccountUpdate,
@@ -51,7 +58,6 @@ export default async function (
     } = await import("o1js");
     const { FungibleToken, FungibleTokenAdmin } =
       await import("mina-fungible-token");
-    const { DefaultConfigs, DefaultModules } = await import("@proto-kit/stack");
 
     const { runtime, protocol } = await loadUserModules();
     const appChain = AppChain.from({
@@ -61,8 +67,8 @@ export default async function (
         ...protocol.settlementModules,
       }),
       Sequencer: Sequencer.from({
-        ...DefaultModules.prismaRedisDatabase(),
-        ...DefaultModules.settlementScript(),
+        Database: InMemoryDatabase,
+        ...scriptModules,
       }),
     });
 
@@ -73,15 +79,8 @@ export default async function (
         ...protocol.settlementModulesConfig,
       },
       Sequencer: {
-        ...DefaultConfigs.prismaRedisDatabase({
-          preset: "development",
-          overrides: {
-            pruneOnStartup: false,
-          },
-        }),
-        ...DefaultConfigs.settlementScript({
-          preset: "development",
-        }),
+        Database: {},
+        ...scriptModulesConfig,
       },
     });
 
@@ -256,10 +255,10 @@ export default async function (
 
     Provable.log("Deployed and initialized settlement contracts", {
       settlement: PrivateKey.fromBase58(
-        process.env.PROTOKIT_SETTLEMENT_CONTRACT_PRIVATE_KEY!
+        getRequiredEnv("PROTOKIT_SETTLEMENT_CONTRACT_PRIVATE_KEY")
       ).toPublicKey(),
       dispatcher: PrivateKey.fromBase58(
-        process.env.PROTOKIT_DISPATCHER_CONTRACT_PRIVATE_KEY!
+        getRequiredEnv("PROTOKIT_DISPATCHER_CONTRACT_PRIVATE_KEY")
       ).toPublicKey(),
     });