+++ packages/protocol/src/prover/transaction/TransactionProver.ts
@@ -408,8 +408,8 @@ export class TransactionProverZkProgrammable extends ZkProgrammable<
 
         proveTransactions: {
           privateInputs: [
-            DynamicRuntimeProof,
-            DynamicRuntimeProof,
+            runtimeProofType,
+            runtimeProofType,
             TransactionProverExecutionData,
             TransactionProverExecutionData,
           ],