+++ packages/sequencer/src/protocol/production/tasks/compile/CircuitCompileTask.ts
@@ -84,8 +84,6 @@ export abstract class CircuitCompileTask<
   public abstract getTargets(): Promise<CompilableModule[]>;
 
   public async compute(input: CompilerTaskParams): Promise<ArtifactRecord> {
-    log.info("Computing VKs");
-
     this.compileRegistry.addArtifactsRaw(input.existingArtifacts);
 
     // We need to initialize the VK tree root if we have it, so that