+++ packages/sequencer/src/settlement/SettlementModule.ts
@@ -16,6 +16,7 @@ import {
   DependencyRecord,
   log,
   dependencyFactory,
+  tryNTimes,
 } from "@proto-kit/common";
 
 import {
@@ -26,6 +27,7 @@ import type { MinaBaseLayer } from "../protocol/baselayer/MinaBaseLayer";
 import { SettleableBatch } from "../storage/model/Batch";
 import { Settlement } from "../storage/model/Settlement";
 import { SettlementStorage } from "../storage/repositories/SettlementStorage";
+import { SettlementInstrumentation } from "../metrics/SettlementInstrumentation";
 
 import { SettlementUtils } from "./utils/SettlementUtils";
 import type { BridgingModule } from "./BridgingModule";
@@ -82,6 +84,9 @@ export class SettlementModule
     return {
       AddressRegistry: {
         useClass: InMemoryAddressRegistry,
+      },
+      SettlementInstrumentation: {
+        useClass: SettlementInstrumentation,
       },
     };
   }
@@ -157,7 +162,12 @@ export class SettlementModule
         : this.parentContainer.dependencyContainer.resolve(
             VanillaSettlementInteraction
           );
-    const settlement = await interaction.settle(batch, options);
+
+    const settlement = await tryNTimes(
+      async () => await interaction.settle(batch, options),
+      3,
+      1000
+    );
 
     await this.settlementStorage.pushSettlement(settlement);